completing the assignment with addition of authentication and authorization

+++ express_authorization/authorization/details_controller.js
@@ -3,7 +3,7 @@ const path = require('path');
 const detailsDbPath = path.join(__dirname, 'details.json');
 const controller_function = (req, res) => {
 	const newDetails = req.body;
-	const api_key = `${newDetails.Name}_${newDetails.password}`;
+	newDetails.api_key = `${newDetails.Name}_${newDetails.password}`;
 	fs.readFile(detailsDbPath, 'utf-8', (err, data) => {
 		if (err) {
 			return res.status(404).json({ error: 'No details found' });
@@ -14,6 +14,12 @@ const controller_function = (req, res) => {
 			newDetails.level_type = 'user';
 		}
 		const oldDetails = JSON.parse(data);
+		const existingUser = oldDetails.find((user) => {
+			return user.Name === newDetails.Name;
+		});
+		if (existingUser) {
+			return res.status(400).json({ massage: 'Existing user' });
+		}
 		const allDetails = [...oldDetails, newDetails];
 		const savedDetails = JSON.stringify(allDetails);
 		fs.writeFile(detailsDbPath, savedDetails, (err) => {
